backend/ingest/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-pubsub==2.19.0
